cleanup: Misc lint/typescript fixes (#7)
* Run prettier

Just gonna sneak this one in

* Add messages file

Adds a messages file to give some type safety for client/server messages. This was one source of `any` types here

* Fix some easy typescript errors

Some any -> unknown and a missing config type

* Swap out custom timestamp scalar any -> string | number

We use a custom timestamp scalar which will generate as `any` but should be `string | number`. Manually switch until we have a reasonable way to generate this (since our graph isnt public)

* Add lint command

In the current absence of an eslint setup, add a temporary `prettier check`