Fix up types for starwars schema
Regenerating the lockfile bumped up the version of graphql which invlidated this type. Really we should switch from the existing `"14.0.2 - 14.2.0 || ^14.3.1 || ^15.0.0"` to something more reasonable :sweat-smile:.

Really we should switch to 15, should be easy once we move off of `@apollographql/graphql-language-service-interface` in favor of `graphql-language-service-interface`. This fork was to include a fix that has since been merged so should be safe.