Switch to graphql-code-generator
graphqlTypes.ts was generated by `apollo codegen` when the VSCode
extension was part of the apollo-tooling repo. When extracted to its own
repo, there was no mechanism checked in  to this repo to regenerate it.

This PR adds a script to regenerate it, and makes that script use
graphql-code-generator rather than `apollo codegen`.