Change dimensions read for timing hints
This shifts to new fields that don't require use to parse out the parent
type and field name manually. (Studio will continue to serve the old
field for a long time so that old extensions continue to work, though it
may stop including real return types after the colon since this code
doesn't need it).

Also add "MS" to a bunch of names to help make clear that vscode-graphql
does not have a bug similar to one recently fixed in Explorer where
units for timing hints got misused.